# src/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Upgrade Journey
schemes:
  # tip: remove http to make production-grade
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    in: header
    name: Authorization
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /tmf-api/agreementManagement/v4/agreement:
    put:
      tags:
        - Plan Details
      summary: Current Contract summary
      operationId: getPlanDetails
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: msisdn
          description: Msisdn of account.
          required: true
          schema:
            $ref: "#/definitions/PlanDetails-Model"
      responses:
        200:
          description: Plan Details retrieved successfully.
        400:
          description: Missing required input parameter.
        404:
          description: No Data found.
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure.
        500:
          description: Technical error occured, service failure.
  /tmf-api/productOfferingQualification/v4/{msisdn}:
    get:
      tags:
        - Eligibility Check
      summary: Returns eligible
      operationId: getUpgradeEligible
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - in: path
          name: msisdn
          required: true
          type: string
          description: Customer MSISDN is input.
      responses:
        200:
          description: Eligibility Check retrieved successfully.
          schema:
            $ref: "#/definitions/EligibilityCheck-Model"
        400:
          description: Missing required input parameter.
        404:
          description: No Data found.
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure.
        500:
          description: Technical error occured, service failure.
  /tmf-api/usageManagement/v4/usage/consumptions:
    get:
      tags:
        - Consumption Usage
      security:
        - Bearer: []
      description: Gets the customers consumption
      operationId: getConsumptionInfo
      parameters:
        - name: msisdn
          in: query
          type: string
          required: true
          description: The msisdn to retrieve the customers consumption
      responses:
        200:
          description: The customers consumption was successfully returned
          schema:
            $ref: "#/definitions/Consumption-Usage"
        "400":
          description: Missing required input parameter
        "424":
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        "500":
          description: Technical error occured, service failure.
  /tmf-api/productCatalogManagement/v4/productSpecification:
    get:
      tags:
        - "Devices"
      operationId: getDeviceModels
      security:
        - Bearer: []
      summary: Returns all device models for a manufacturer
      consumes:
        - application/json
      parameters:
        - in: "query"
          name: brandName
          type: array
          items:
            type: string
          required: false
      responses:
        200:
          description: All device models retrieved successfully
          schema:
            $ref: "#/definitions/DeviceModel-Model"
        400:
          description: Missing required input parameter
        404:
          description: No device models found
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        500:
          description: Technical error occured, service failure
...
  /tmf-api/party/v4/individual/validate:
    get:
      tags:
        - "Call Prevetting"
      operationId: callPreVetting
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - in: query
          name: idNumber
          type: string
          required: true
        - in: query
          name: idType
          type: string
          required: true
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/CallPrevetting-Model"
        400:
          description: Missing required input parameter
        401:
          description: Not authorized
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        500:
          description: Technical error occured, service failure.
  /tmf-api/productOfferingQualification/v4/customers/{idNumber}/information:
    get:
      tags:
        - "Affordability Check"
      operationId: getAffordabilityCheck
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - in: path
          name: idNumber
          type: string
          required: true
        - in: query
          name: field
          type: string
          required: true
        - in: query
          name: idType
          type: string
          required: true
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/AffordabilityCheck"
        400:
          description: Missing required input parameter
        401:
          description: Not authorized
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        500:
          description: Technical error occured, service failure.
...
  /tmf-api/productOrderingManagement/v4/productOrder:
    post:
      tags:
        - "COFQ Create Order"
      operationId: orderCreation
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: upgradeModel
          description: upgradeModel
          required: true
          schema:
            $ref: "#/definitions/CofqSubmitUpgrade"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/CofqSubmitUpgrade"
        400:
          description: Missing required input parameter
        401:
          description: Not authorized
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        500:
          description: Technical error occured, service failure.
  /tmf-api/productOrderingManagement/v4/productOrder/submitupgrade:
    put:
      tags:
        - "CQFD Submit Upgrade"
      operationId: submitUpgrade
      security:
        - Bearer: []
      consumes:
        - application/json
      parameters:
        - in: body
          name: upgradeModel
          description: upgradeModel
          required: true
          schema:
            $ref: "#/definitions/SubmitUpgrade"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SubmitUpgrade"
        400:
          description: Missing required input parameter
        401:
          description: Not authorized
        424:
          description: Upstream call failed with a 500. Response will be identical to upstream failure
        500:
          description: Technical error occured, service failure.
...
definitions:
...
  PlanDetails-Model:
    type: object
    properties:
      agreementItem:
        type: array
        items:
          type: object
          properties:
            product:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
            termOrCondition:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                  validFor:
                    type: object
                    properties:
                      endDateTime:
                        type: string
                      startDateTime:
                        type: string
      characteristic:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            value:
              type: string

        # Get EligibilityCheck
  EligibilityCheck-Model:
    type: object
    properties:
      productOfferingQualificationDate:
        type: string
      qualificationResult:
        type: string
      productOfferingQualificationItem:
        type: array
        items:
          type: object
          properties:
            eligibilityUnavailabilityReason:
              type: array
              items:
                type: object
                properties:
                  code:
                    type: string
            note:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: string
            product:
              type: object
              properties:
                productCharacteristic:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      value:
                        type: string
      relatedParty:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            name:
              type: string

        # Get Device Models
  DeviceModel-Model:
    type: object
    properties:
      brand:
        type: string
      description:
        type: string
      productSpecCharacteristic:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            valueType:
              type: string
            productSpecCharacteristicValue:
              type: object
              properties:
                value:
                  type: string

    # Get Manufacturer List
...
  CallPrevetting-Model:
    type: object
    properties:
      partyCharacteristic:
        type: array
        items:
          type: object
          properties:
            name:
              type: string
            value:
              type: string

    # Get Consumption Usage
  Consumption-Usage:
    type: object
    properties:
      usageCharacteristic:
        type: array
        items:
          type: string
...
  AffordabilityCheck:
    type: object
    properties:
      productOfferingQualificationItem:
        type: array
        items:
          type: object
          properties:
            product:
              type: object
              properties:
                description:
                  type: string
                productCharacteristic:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      value:
                        type: string
  CofqSubmitUpgrade:
    type: object
    properties:
      channel:
        type: string
      msisdn:
        type: string
      imei:
        type: string
      paymentFlag:
        type: string
      quoteNumber:
        type: string
      submitFlag:
        type: string
      userId:
        type: string
    title: CofqSubmitUpgrade
  SubmitUpgrade:
    type: object
    properties:
      checkoutAuthToken:
        type: string
      channel:
        type: string
      dealsheetNumber:
        type: string
      email:
        type: string
      msisdn:
        type: string
      storeIdentifier:
        type: string
      imei:
        type: string
      deliveryAddress:
        $ref: "#/definitions/deliveryAddress"
    title: SubmitUpgrade
  deliveryAddress:
    type: object
    properties:
      addressType:
        type: string
      buildingName:
        type: string
      streetNo:
        type: string
      streetName:
        type: string
      suburb:
        type: string
      postalCode:
        type: string
      unitNumber:
        type: string
      city:
        type: string
      province:
        type: string
    title: deliveryAddress